requirements.txt
absl-py==0.7.1
bandit==1.6.2
click==7.0
cloudpickle==1.2.1
flake8==3.7.8
gin-config==0.2.0
gym==0.13.1
ipdb==0.12.2
isort==4.3.21
matplotlib==3.1.1
mccabe==0.6.1
mypy==0.720
networkx==2.3
numpy==1.16.4
psutil==5.4.8
pytest==5.0.1
pyparsing==2.4.2
pyyaml==5.1.1
pytest==5.0.1
snakeviz==2.0.1
tensorboard==1.14.0
tensorboardX==1.8
tensorflow==1.14.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.2
ujson==1.35
vulture==1.0
wandb==0.8.9
